Refactor test fixtures
I'm putting common fixtures in tests/conftest.py (a canonical module) so
that all tests can use them, regardless of the test's module.
